MSDKUI-1722: [Android] Component app findings GuidanceSpeed Limit -- it is possible now to show background with no text GuidanceSpeedLimit -- Removed default padding and margins

+++ MSDKUIKit/MSDKUILib/src/main/java/com/here/msdkui/guidance/GuidanceSpeedPresenter.java
@@ -89,7 +89,7 @@ public class GuidanceSpeedPresenter extends BaseGuidancePresenter {
         final double speed = geoPosition != null && geoPosition.isValid() && geoPosition.getSpeed() != GeoPosition.UNKNOWN ?
                 geoPosition.getSpeed() : -1;
         if (speed >= 0) {
-            final GuidanceSpeedData data = new GuidanceSpeedData(speed, speedLimit);
+            final GuidanceSpeedData data = new GuidanceSpeedData(speed, (double) speedLimit);
             notifyDataChanged(data);
         } else {
             notifyDataChanged(null);
